+++ swing/framework-ui/src/main/java/is/codion/swing/framework/ui/EntityEditComponentPanel.java
@@ -7,6 +7,7 @@ import is.codion.common.Configuration;
 import is.codion.common.property.PropertyValue;
 import is.codion.common.state.State;
 import is.codion.common.value.Value;
+import is.codion.common.value.ValueSet;
 import is.codion.framework.domain.entity.Entities;
 import is.codion.framework.domain.entity.Entity;
 import is.codion.framework.domain.entity.attribute.Attribute;
@@ -118,7 +119,7 @@ public class EntityEditComponentPanel extends JPanel {
   private final EntityComponents entityComponents;
   private final Map<Attribute<?>, Value<JComponent>> components = new HashMap<>();
   private final Map<Attribute<?>, ComponentBuilder<?, ?, ?>> componentBuilders = new HashMap<>();
-  private final Set<Attribute<?>> excludeFromSelection = new HashSet<>();
+  private final ValueSet<Attribute<?>> excludeComponentFromSelection = ValueSet.valueSet();
   private final Value<JComponent> focusedInputComponent = Value.value();
 
   private final Value<JComponent> initialFocusComponent = Value.value();
@@ -151,6 +152,8 @@ public class EntityEditComponentPanel extends JPanel {
     if (!editModel.entityType().equals(entityComponents.entityDefinition().entityType())) {
       throw new IllegalArgumentException("Entity type mismatch: " + editModel.entityType() + ", " + entityComponents.entityDefinition().entityType());
     }
+    excludeComponentFromSelection.addValidator(attributes -> attributes.forEach(attribute ->
+            requireNonNull(attribute, "Excluded component attribute may not be null")));
     addFocusedComponentListener();
   }
 
@@ -268,7 +271,7 @@ public class EntityEditComponentPanel extends JPanel {
    * Displays a dialog allowing the user the select an input component which should receive the keyboard focus.
    * If only one input component is available then that component is selected automatically.
    * If no component is available, f.ex. when the panel is not visible, this method does nothing.
-   * @see #excludeComponentsFromSelection(Attribute[])
+   * @see #excludeComponentFromSelection()
    * @see #requestComponentFocus(Attribute)
    */
   public final void selectInputComponent() {
@@ -291,25 +294,23 @@ public class EntityEditComponentPanel extends JPanel {
    * @return an unmodifiable view of the attributes to present when selecting an input component in this panel,
    * this returns all (non-excluded) attributes that have an associated component in this panel
    * that is enabled, displayable, visible and focusable.
-   * @see #excludeComponentsFromSelection(Attribute[])
+   * @see #excludeComponentFromSelection()
    * @see #component(Attribute)
    */
   public final Collection<Attribute<?>> selectComponentAttributes() {
     return components.keySet().stream()
-            .filter(attribute -> !excludeFromSelection.contains(attribute))
+            .filter(attribute -> !excludeComponentFromSelection.contains(attribute))
             .filter(attribute -> componentSelectable(component(attribute).get()))
             .collect(collectingAndThen(Collectors.toList(), Collections::unmodifiableCollection));
   }
 
   /**
-   * Specifies that the given attributes should be excluded when presenting a component selection list.
-   * @param attributes the attributes to exclude from selection
+   * Specifies the attributes that should be excluded when presenting a component selection list.
+   * @return the {@link ValueSet} specifying attributes that should be excluded from component selection
    * @see #selectInputComponent()
    */
-  public final void excludeComponentsFromSelection(Attribute<?>... attributes) {
-    for (Attribute<?> attribute : requireNonNull(attributes)) {
-      excludeFromSelection.add(requireNonNull(attribute));
-    }
+  public final ValueSet<Attribute<?>> excludeComponentFromSelection() {
+    return excludeComponentFromSelection;
   }
 
   /**